Add code action for reporting issue + refactor
